requirements.txt
APScheduler==3.0.1
Flask==0.10.1
Flask-Moment==0.4.0
Flask-SQLAlchemy==2.0
Flask-WTF==0.10.3
Jinja2==2.7.3
MarkupSafe==0.23
SQLAlchemy==0.9.8
SQLAlchemy-Utils==0.29.2
Tempita==0.5.2
WTForms==2.0.1
Werkzeug==0.9.6
argparse==1.2.1
coinkite-api==1.0.0
decorator==3.4.0
futures==2.2.0
gunicorn==19.1.1
infinity==1.3
intervals==0.3.1
itsdangerous==0.24
pbr==0.10.7
psycopg2==2.5.4
python-dateutil==2.4.0
pytz==2014.10
requests==2.5.1
simplejson==3.6.5
six==1.9.0
sqlalchemy-migrate==0.9.4
sqlparse==0.1.14
tzlocal==1.1.2
validators==0.7
wsgiref==0.1.2
